# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barberia",
        "description": "Servicio de API tipo SOA para implementación en una barberia",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/detalleproducto/nuevo": {
            "post": {
                "tags": [
                    "DetalleProducto"
                ],
                "summary": "Crear un nuevo detalle de producto",
                "operationId": "477db8c710ffa5881eeb79c147f7170d",
                "requestBody": {
                    "description": "Datos del detalle de producto para crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdProducto": {
                                        "description": "ID del producto",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "IdUsuario": {
                                        "description": "ID del usuario",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "CantidadProductos": {
                                        "description": "Cantidad de productos",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "PrecioUnitario": {
                                        "description": "Precio unitario del producto",
                                        "type": "number",
                                        "example": "10.99"
                                    },
                                    "MontoTotal": {
                                        "description": "Monto total del detalle",
                                        "type": "number",
                                        "example": "54.95"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha creado el detalle de producto correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al crear el detalle de producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al crear el detalle de producto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/detalleproducto/id/editar": {
            "put": {
                "tags": [
                    "DetalleProducto"
                ],
                "summary": "Actualizar un detalle de producto específico",
                "operationId": "ab6ad6ee03889d3e40fe7e1871977456",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del detalle de producto para actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdProducto": {
                                        "description": "ID del producto",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "IdUsuario": {
                                        "description": "ID del usuario",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "CantidadProductos": {
                                        "description": "Cantidad de productos",
                                        "type": "integer",
                                        "example": "5"
                                    },
                                    "PrecioUnitario": {
                                        "description": "Precio unitario del producto",
                                        "type": "number",
                                        "example": "10.99"
                                    },
                                    "MontoTotal": {
                                        "description": "Monto total del detalle",
                                        "type": "number",
                                        "example": "54.95"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El detalle de producto ha sido actualizado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al actualizar el detalle de producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al actualizar el detalle de producto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/detalleproducto/id": {
            "get": {
                "tags": [
                    "DetalleProducto"
                ],
                "summary": "Mostrar un detalle de producto específico",
                "operationId": "9c5c0db7cfceb0959f9262db669e30b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de producto a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "IdDetalleProducto": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "IdProducto": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "IdUsuario": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "CantidadProductos": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "PrecioUnitario": {
                                            "type": "number",
                                            "example": "10.99"
                                        },
                                        "MontoTotal": {
                                            "type": "number",
                                            "example": "54.95"
                                        },
                                        "created_at": {
                                            "description": "Fecha de creación del detalle de producto",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Fecha de última actualización del detalle de producto",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al mostrar el detalle de producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al mostrar el detalle de producto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/detalleproducto/id/eliminar": {
            "delete": {
                "tags": [
                    "DetalleProducto"
                ],
                "summary": "Eliminar un detalle de producto específico",
                "operationId": "21da91167bfc0002099f35b7e18d8f8d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del detalle de producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El detalle de producto ha sido eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al eliminar el detalle de producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al eliminar el detalle de producto."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}